agnoKnowledge/pyproject.toml
[project]
name = "agnoknowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.8",
    "chonkie[openai]>=1.4.2",
    "lancedb>=0.25.3",
    "openai>=2.9.0",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
]
